refactor: PublicWorkbook => PublicSearchResultPage
