+++ frontend/src/Router.tsx
@@ -20,7 +20,7 @@ import {
   MainPage,
   PublicCardsPage,
   PublicSearchPage,
-  PublicWorkbookPage,
+  PublicSearchResultPage,
   QuizPage,
   QuizResultPage,
   QuizSettingPage,
@@ -84,8 +84,8 @@ const routes = [
     isPublic: false,
   },
   {
-    path: ROUTE.PUBLIC_WORKBOOK.PATH,
-    component: <PublicWorkbookPage />,
+    path: ROUTE.PUBLIC_SEARCH_RESULT.PATH,
+    component: <PublicSearchResultPage />,
     isPublic: false,
   },
   {
